Moved SwiftyTextTable extension into RulesCommand.
